+++ CHANGELOG.md
@@ -53,6 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `jj git fetch` now automatically prints new remote branches and tags by default.
 
+* Added completions for [Nushell](https://nushell.sh) to `jj util completion`
+
 ### Fixed bugs
 
 * On Windows, symlinks in the repo are now materialized as regular files in the